Refactored desktop application
